refactor: del AStar- og MST-visualiseringene i mindre, separerte filer
Trekker ut delt useVisualizationEnvironment-hook, samt scene/settings/colors/
render for A* og scene/settings for MST. Komponentene beholder kun
animasjonslivssyklusen.
